Move test files to test dir
