+++ client-react/src/pages/App/App.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { screen } from '@testing-library/react';
-import { renderWithThemeAndRedux } from '../../test/test-utils';
+import { renderWithThemeAndRedux } from '../../test/utils';
 import App from './index';
 import { createStore } from '../../store/store';
