+++ client-react/src/pages/Register/Register.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { screen } from '@testing-library/react';
-import { renderWithThemeAndRedux } from '../../test/test-utils';
+import { renderWithThemeAndRedux } from '../../test/utils';
 import userEvent from '@testing-library/user-event';
 import Register from './index';
 import chat from '../../shared/chat';
